Add Requirements on XY Noise.
This addition was done without cleaning up collateral damage.
For example, there are some missing references, a pdf figure
referenced in the Tex files that doesn't have a source SVG file
or script, and in general I didn't make sure the narration is clean.
That needs to happen in another commit.